# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Toto Game Rule Service API in Go using Gin framework.",
        "title": "Toto Game Rule Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:9090",
    "paths": {
        "/": {
            "get": {
                "description": "Responds with \"Hello world\" message.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "root"
                ],
                "summary": "Test root",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.RootResponse"
                        }
                    }
                }
            }
        },
        "/rules": {
            "get": {
                "description": "Responds with the list of all rules as JSON.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rules"
                ],
                "summary": "Get rules array",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.RulesResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Responds with the rule created of as JSON.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rules"
                ],
                "summary": "Add Rule",
                "parameters": [
                    {
                        "description": "Add rule",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Rule"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/responses.RuleResponse"
                        }
                    }
                }
            }
        },
        "/rules/{id}": {
            "get": {
                "description": "Responds with a rule as JSON",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rules"
                ],
                "summary": "Get rule by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Rule ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.RuleResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Responds with a message as JSON",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rules"
                ],
                "summary": "Delete rule by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Rule ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.RuleGeneralResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Answer": {
            "type": "object",
            "required": [
                "text"
            ],
            "properties": {
                "order": {
                    "type": "integer",
                    "example": 1
                },
                "text": {
                    "type": "string",
                    "example": "12 o'clock"
                }
            }
        },
        "models.Rule": {
            "type": "object",
            "required": [
                "answers",
                "question"
            ],
            "properties": {
                "answers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Answer"
                    }
                },
                "id": {
                    "type": "string",
                    "example": "634f787edc90bf2b7c868153"
                },
                "order": {
                    "type": "integer",
                    "example": 1
                },
                "question": {
                    "type": "string",
                    "example": "What time is it?"
                }
            }
        },
        "responses.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "the error"
                },
                "message": {
                    "type": "string",
                    "example": "fail"
                },
                "status": {
                    "type": "integer",
                    "example": 500
                }
            }
        },
        "responses.RootResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "success"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
        "responses.RuleGeneralResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "success"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
        "responses.RuleResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.Rule"
                },
                "message": {
                    "type": "string",
                    "example": "success"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
        "responses.RulesResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Rule"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "success"
                },
                "results": {
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        }
    }
}